+++ src/Neo.CLI/CLI/MainService.cs
@@ -161,79 +161,6 @@ namespace Neo.CLI
             CurrentWallet = wallet;
         }
 
-        private IEnumerable<Block> GetBlocks(Stream stream, bool read_start = false)
-        {
-            using BinaryReader r = new BinaryReader(stream);
-            uint start = read_start ? r.ReadUInt32() : 0;
-            uint count = r.ReadUInt32();
-            uint end = start + count - 1;
-            uint currentHeight = NativeContract.Ledger.CurrentIndex(NeoSystem.StoreView);
-            if (end <= currentHeight) yield break;
-            for (uint height = start; height <= end; height++)
-            {
-                var size = r.ReadInt32();
-                if (size > Message.PayloadMaxSize)
-                    throw new ArgumentException($"Block {height} exceeds the maximum allowed size");
-
-                byte[] array = r.ReadBytes(size);
-                if (height > currentHeight)
-                {
-                    Block block = array.AsSerializable<Block>();
-                    yield return block;
-                }
-            }
-        }
-
-        private IEnumerable<Block> GetBlocksFromFile()
-        {
-            const string pathAcc = "chain.acc";
-            if (File.Exists(pathAcc))
-                using (FileStream fs = new(pathAcc, FileMode.Open, FileAccess.Read, FileShare.Read))
-                    foreach (var block in GetBlocks(fs))
-                        yield return block;
-
-            const string pathAccZip = pathAcc + ".zip";
-            if (File.Exists(pathAccZip))
-                using (FileStream fs = new(pathAccZip, FileMode.Open, FileAccess.Read, FileShare.Read))
-                using (ZipArchive zip = new(fs, ZipArchiveMode.Read))
-                using (Stream? zs = zip.GetEntry(pathAcc)?.Open())
-                {
-                    if (zs is not null)
-                    {
-                        foreach (var block in GetBlocks(zs))
-                            yield return block;
-                    }
-                }
-
-            var paths = Directory.EnumerateFiles(".", "chain.*.acc", SearchOption.TopDirectoryOnly).Concat(Directory.EnumerateFiles(".", "chain.*.acc.zip", SearchOption.TopDirectoryOnly)).Select(p => new
-            {
-                FileName = Path.GetFileName(p),
-                Start = uint.Parse(Regex.Match(p, @"\d+").Value),
-                IsCompressed = p.EndsWith(".zip")
-            }).OrderBy(p => p.Start);
-
-            uint height = NativeContract.Ledger.CurrentIndex(NeoSystem.StoreView);
-            foreach (var path in paths)
-            {
-                if (path.Start > height + 1) break;
-                if (path.IsCompressed)
-                    using (FileStream fs = new(path.FileName, FileMode.Open, FileAccess.Read, FileShare.Read))
-                    using (ZipArchive zip = new(fs, ZipArchiveMode.Read))
-                    using (Stream? zs = zip.GetEntry(Path.GetFileNameWithoutExtension(path.FileName))?.Open())
-                    {
-                        if (zs is not null)
-                        {
-                            foreach (var block in GetBlocks(zs, true))
-                                yield return block;
-                        }
-                    }
-                else
-                    using (FileStream fs = new(path.FileName, FileMode.Open, FileAccess.Read, FileShare.Read))
-                        foreach (var block in GetBlocks(fs, true))
-                            yield return block;
-            }
-        }
-
         private bool NoWallet()
         {
             if (CurrentWallet != null) return false;
@@ -375,7 +302,7 @@ namespace Neo.CLI
             bool verifyImport = !(options.NoVerify ?? false);
 
             Utility.LogLevel = options.Verbose;
-            ProtocolSettings protocol = ProtocolSettings.Load("config.json");
+            var protocol = ProtocolSettings.Load("config.json");
             CustomProtocolSettings(options, protocol);
             CustomApplicationSettings(options, Settings.Default);
             try
@@ -439,25 +366,8 @@ namespace Neo.CLI
                 RegisterCommand(plugin, plugin.Name);
             }
 
-            using (IEnumerator<Block> blocksBeingImported = GetBlocksFromFile().GetEnumerator())
-            {
-                while (true)
-                {
-                    List<Block> blocksToImport = new List<Block>();
-                    for (int i = 0; i < 10; i++)
-                    {
-                        if (!blocksBeingImported.MoveNext()) break;
-                        blocksToImport.Add(blocksBeingImported.Current);
-                    }
-                    if (blocksToImport.Count == 0) break;
-                    await NeoSystem.Blockchain.Ask<Blockchain.ImportCompleted>(new Blockchain.Import
-                    {
-                        Blocks = blocksToImport,
-                        Verify = verifyImport
-                    });
-                    if (NeoSystem is null) return;
-                }
-            }
+            await ImportBlocksFromFile(verifyImport);
+
             NeoSystem.StartNode(new ChannelsConfig
             {
                 Tcp = new IPEndPoint(IPAddress.Any, Settings.Default.P2P.Port),
@@ -513,52 +423,6 @@ namespace Neo.CLI
         {
             Dispose_Logger();
             Interlocked.Exchange(ref _neoSystem, null)?.Dispose();
-        }
-
-        private void WriteBlocks(uint start, uint count, string path, bool writeStart)
-        {
-            uint end = start + count - 1;
-            using FileStream fs = new FileStream(path, FileMode.OpenOrCreate, FileAccess.ReadWrite, FileShare.None, 4096, FileOptions.WriteThrough);
-            if (fs.Length > 0)
-            {
-                byte[] buffer = new byte[sizeof(uint)];
-                if (writeStart)
-                {
-                    fs.Seek(sizeof(uint), SeekOrigin.Begin);
-                    fs.ReadExactly(buffer);
-                    start += BitConverter.ToUInt32(buffer, 0);
-                    fs.Seek(sizeof(uint), SeekOrigin.Begin);
-                }
-                else
-                {
-                    fs.ReadExactly(buffer);
-                    start = BitConverter.ToUInt32(buffer, 0);
-                    fs.Seek(0, SeekOrigin.Begin);
-                }
-            }
-            else
-            {
-                if (writeStart)
-                {
-                    fs.Write(BitConverter.GetBytes(start), 0, sizeof(uint));
-                }
-            }
-            if (start <= end)
-                fs.Write(BitConverter.GetBytes(count), 0, sizeof(uint));
-            fs.Seek(0, SeekOrigin.End);
-            Console.WriteLine("Export block from " + start + " to " + end);
-
-            using (var percent = new ConsolePercent(start, end))
-            {
-                for (uint i = start; i <= end; i++)
-                {
-                    Block block = NativeContract.Ledger.GetBlock(NeoSystem.StoreView, i);
-                    byte[] array = block.ToArray();
-                    fs.Write(BitConverter.GetBytes(array.Length), 0, sizeof(int));
-                    fs.Write(array, 0, array.Length);
-                    percent.Value = i;
-                }
-            }
         }
 
         private static void WriteLineWithoutFlicker(string message = "", int maxWidth = 80)
